Split events, create folders
