+++ CheckersApplication/CheckersApplication/Events/Camera.cs
@@ -1,0 +1,48 @@
+﻿using System;
+using System.Windows;
+using System.Windows.Interop;
+
+using Emgu.CV;
+
+using DirectShowLib;
+
+namespace CheckersApplication
+{
+    public partial class MainWindow : Window
+    {
+        private void BT_Start_Click(object sender, RoutedEventArgs e)
+        {
+            ChangeBtnStartStop();
+            if (CB_DefaultCamera.IsChecked == true && CO_Cameras.Items.Count > 0)
+                camera = new Camera(Convert.ToString(CO_Cameras.SelectedIndex));
+            else
+                camera = new Camera(TB_CameraSource.Text);
+            CameraShow();
+            CO_Cameras.IsEnabled = false;
+        }
+
+        private void BT_Stop_Click(object sender, RoutedEventArgs e)
+        {
+            ChangeBtnStartStop();
+            ComponentDispatcher.ThreadIdle -= new EventHandler(updateFrames);
+            camera.capture.Stop();
+            camera.capture.Dispose();
+            CO_Cameras.IsEnabled = true;
+        }
+
+        private void CB_DefaultCamera_Click(object sender, RoutedEventArgs e)
+        {
+            CO_Cameras.IsEnabled = !CO_Cameras.IsEnabled;
+            TB_CameraSource.IsEnabled = !(bool)CB_DefaultCamera.IsChecked;
+            if (!TB_CameraSource.IsEnabled)
+            {
+                TB_CameraSource.Text = "URL streamu, ID kamery, lub plik wideo,\nnp. (http://IP:PORT/mjpegfeed)";
+                TB_CameraSource.Foreground = System.Windows.Media.Brushes.Gray;
+            }
+        }
+
+       
+
+
+    }
+}
