+++ CheckersApplication/CheckersApplication/Helper/ToBitmapConverter.cs
