+++ CheckersApplication/CheckersApplication/MainWindow.xaml.cs
@@ -109,62 +109,7 @@ namespace CheckersApplication
             ComponentDispatcher.ThreadIdle += new System.EventHandler(updateFrames);
         }
 
-        private void BT_Start_Click(object sender, RoutedEventArgs e)
-        {
-            ChangeBtnStartStop();
-            if (CB_DefaultCamera.IsChecked == true && CO_Cameras.Items.Count > 0)
-                camera = new Camera(Convert.ToString(CO_Cameras.SelectedIndex));
-            else
-                camera = new Camera(TB_CameraSource.Text);
-            CameraShow();
-            CO_Cameras.IsEnabled = false;
-        }
-
-        private void BT_Stop_Click(object sender, RoutedEventArgs e)
-        {
-            ChangeBtnStartStop();
-            ComponentDispatcher.ThreadIdle -= new EventHandler(updateFrames);
-            camera.capture.Stop();
-            camera.capture.Dispose();
-            CO_Cameras.IsEnabled = true;
-        }
-
-        private void CB_DefaultCamera_Click(object sender, RoutedEventArgs e)
-        {
-            CO_Cameras.IsEnabled = !CO_Cameras.IsEnabled;
-            TB_CameraSource.IsEnabled = !(bool)CB_DefaultCamera.IsChecked;
-            if (!TB_CameraSource.IsEnabled)
-            {
-                TB_CameraSource.Text = "URL streamu, ID kamery, lub plik wideo,\nnp. (http://IP:PORT/mjpegfeed)";
-                TB_CameraSource.Foreground = System.Windows.Media.Brushes.Gray;
-            }
-        }
-
-        private void CB_OpenCL_Click(object sender, RoutedEventArgs e)
-        {
-            CvInvoke.UseOpenCL = !CvInvoke.UseOpenCL;
-        }
-
-        private void BT_ImageTest_Click(object sender, RoutedEventArgs e)
-        {
-            OpenFileDialog openFileDialog1 = new OpenFileDialog();
-            openFileDialog1.Filter = "Image files (*.jpg, *.png) | *.jpg; *.png";
-            openFileDialog1.RestoreDirectory = true;
-            DialogResult result = openFileDialog1.ShowDialog();
-            string path = String.Empty;
-            if (result == System.Windows.Forms.DialogResult.OK)
-            {
-                try
-                {
-                    fileName = openFileDialog1.FileName;
-                    Detect(openFileDialog1.FileName);
-                }
-                catch (Exception ex)
-                {
-                    System.Windows.MessageBox.Show("Komunikat błędu: " + ex.Message);
-                }
-            }
-        }
+        
 
         private void TB_CameraSource_GotFocus(object sender, RoutedEventArgs e)
         {
@@ -395,267 +340,12 @@ namespace CheckersApplication
         }
         */
 
-        private void CB_AutoDetectColors_Click(object sender, RoutedEventArgs e)
-        {
-            if (CB_AutoDetectColors.IsChecked == true)
-            {
-                RS_Slider1R.IsEnabled = false;
-                RS_Slider2R.IsEnabled = false;
-                RS_Slider1G.IsEnabled = false;
-                RS_Slider2G.IsEnabled = false;
-                RS_Slider1B.IsEnabled = false;
-                RS_Slider2B.IsEnabled = false;
-            }
-            else
-            {
-                RS_Slider1R.IsEnabled = true;
-                RS_Slider2R.IsEnabled = true;
-                RS_Slider1G.IsEnabled = true;
-                RS_Slider2G.IsEnabled = true;
-                RS_Slider1B.IsEnabled = true;
-                RS_Slider2B.IsEnabled = true;
-            }
-        }
+        
 
-        private void RangeSlider_Player1_ValueChanged(object sender, RoutedEventArgs e)
-        {
-            if (ifInitComponent)
-            {
-                LB_Range1R.Content = RS_Slider1R.LowerValue.ToString() + " - " + RS_Slider1R.HigherValue.ToString();
-                LB_Range1G.Content = RS_Slider1G.LowerValue.ToString() + " - " + RS_Slider1G.HigherValue.ToString();
-                LB_Range1B.Content = RS_Slider1B.LowerValue.ToString() + " - " + RS_Slider1B.HigherValue.ToString();
+        
 
-                if (CB_AutoDetectColors.IsChecked == false)
-                { 
-                    System.Windows.Media.Color minColor = System.Windows.Media.Color.FromRgb((byte)RS_Slider1R.LowerValue, (byte)RS_Slider1G.LowerValue, (byte)RS_Slider1B.LowerValue);
-                    CV_Player1Color_Min.Background = new System.Windows.Media.SolidColorBrush(minColor);
+        
 
-                    System.Windows.Media.Color maxColor = System.Windows.Media.Color.FromRgb((byte)RS_Slider1R.HigherValue, (byte)RS_Slider1G.HigherValue, (byte)RS_Slider1B.HigherValue);
-                    CV_Player1Color_Max.Background = new System.Windows.Media.SolidColorBrush(maxColor);
-
-                    System.Windows.Application.Current.Resources["Player1Color"] = CV_Player1Color_Max.Background;
-                }
-            }
-        }
-
-        private void RangeSlider_Player2_ValueChanged(object sender, RoutedEventArgs e)
-        {
-            if (ifInitComponent)
-            {
-                LB_Range2R.Content = RS_Slider2R.LowerValue.ToString() + " - " + RS_Slider2R.HigherValue.ToString();
-                LB_Range2G.Content = RS_Slider2G.LowerValue.ToString() + " - " + RS_Slider2G.HigherValue.ToString();
-                LB_Range2B.Content = RS_Slider2B.LowerValue.ToString() + " - " + RS_Slider2B.HigherValue.ToString();
-
-                if (CB_AutoDetectColors.IsChecked == false)
-                {
-                    System.Windows.Media.Color minColor = System.Windows.Media.Color.FromRgb((byte)RS_Slider2R.LowerValue, (byte)RS_Slider2G.LowerValue, (byte)RS_Slider2B.LowerValue);
-                    CV_Player2Color_Min.Background = new System.Windows.Media.SolidColorBrush(minColor);
-
-                    System.Windows.Media.Color maxColor = System.Windows.Media.Color.FromRgb((byte)RS_Slider2R.HigherValue, (byte)RS_Slider2G.HigherValue, (byte)RS_Slider2B.HigherValue);
-                    CV_Player2Color_Max.Background = new System.Windows.Media.SolidColorBrush(maxColor);
-
-                    System.Windows.Application.Current.Resources["Player2Color"] = CV_Player2Color_Max.Background;
-                }
-            }
-        }
-
-        private void BT_SaveMove_Click(object sender, RoutedEventArgs e)
-        {
-            if (currentMove == 0)
-            {
-                currentMove++;
-                shownMove++;
-                TB_MoveNr.Text = "Nr ruchu: " + (shownMove).ToString() + " (bieżący)";
-                List<CheckersPiece> pieces = new List<CheckersPiece>();
-                foreach (var p in chessBoardState.piecesObservable)
-                    pieces.Add(p.Copy());
-
-
-
-
-                chessBoardState.history.Add(pieces);
-            }
-            else if (currentMove % 2 == 1)
-            {
-                goodMove = false;
-                ChessField[,] board = ChessField.GetEmptyFields();
-                ChessField[,] board2 = ChessField.GetEmptyFields();
-                List<CheckersPiece> pieces = new List<CheckersPiece>();
-                pieces = chessBoardState.history[chessBoardState.history.Count - 1];
-
-                foreach (var i in pieces)
-                {
-                    //if (i.Player == Player.BlackMen)
-                    //    board[i.Pos.X / 60, i.Pos.Y / 60].Value = (int)Player.BlackMen;
-                    //if (i.Player == Player.WhiteMen)
-                    //    board[i.Pos.X / 60, i.Pos.Y / 60].Value = (int)Player.WhiteMen;
-                    board[i.Pos.Y / 60, i.Pos.X / 60].Value = (int)i.Player;
-                }
-                //for (int i = 0; i < 8; i++)
-                //{
-                //    for (int j = 0; j < 8; j++)
-                //        File.AppendAllText("log13", (board[i, j].Value).ToString());
-                //    File.AppendAllText("log13", "\r\n");
-                //}
-
-                List<CheckersPiece> pieces2 = new List<CheckersPiece>();
-                foreach (var p in chessBoardState.piecesObservable)
-                    pieces2.Add(p.Copy());
-
-                foreach (var i in pieces2)
-                {
-                    //if (i.Player == Player.BlackMen)
-                    //    board[i.Pos.X / 60, i.Pos.Y / 60].Value = (int)Player.BlackMen;
-                    //if (i.Player == Player.WhiteMen)
-                    //    board[i.Pos.X / 60, i.Pos.Y / 60].Value = (int)Player.WhiteMen;
-                    board2[i.Pos.Y / 60, i.Pos.X / 60].Value = (int)i.Player;
-                }
-
-                buffer_move_white = MovesJumps.RunWhite(board);
-
-                //for (int i = 0; i < 8; i++)
-                //{
-                //    for (int j = 0; j < 8; j++)
-                //        Console.Write(board2[i, j].Value);
-                //    Console.WriteLine();
-                //}
-
-                foreach (var v in buffer_move_white)
-                {
-                    if (same(v, board2))
-                        goodMove = true;
-                }
-
-                if (goodMove == false)
-                    System.Windows.MessageBox.Show("Niepoprawny ruch.");
-                else
-                {
-
-                    currentMove++;
-                    shownMove++;
-                    TB_MoveNr.Text = "Nr ruchu: " + (shownMove).ToString() + " (bieżący)";
-                    chessBoardState.history.Add(pieces2);
-                }
-            }
-            else
-            {
-                goodMove = false;
-                ChessField[,] board = ChessField.GetEmptyFields();
-                ChessField[,] board2 = ChessField.GetEmptyFields();
-                List<CheckersPiece> pieces = new List<CheckersPiece>();
-                pieces = chessBoardState.history[chessBoardState.history.Count - 1];
-
-                foreach (var i in pieces)
-                {
-                    //if (i.Player == Player.BlackMen)
-                    //    board[i.Pos.X / 60, i.Pos.Y / 60].Value = (int)Player.BlackMen;
-                    //if (i.Player == Player.WhiteMen)
-                    //    board[i.Pos.X / 60, i.Pos.Y / 60].Value = (int)Player.WhiteMen;
-                    board[i.Pos.Y / 60, i.Pos.X / 60].Value = (int)i.Player;
-                }
-                //for (int i = 0; i < 8; i++)
-                //{
-                //    for (int j = 0; j < 8; j++)
-                //        File.AppendAllText("log13", (board[i, j].Value).ToString());
-                //    File.AppendAllText("log13", "\r\n");
-                //}
-
-                List<CheckersPiece> pieces2 = new List<CheckersPiece>();
-                foreach (var p in chessBoardState.piecesObservable)
-                    pieces2.Add(p.Copy());
-
-                foreach (var i in pieces2)
-                {
-                    //if (i.Player == Player.BlackMen)
-                    //    board[i.Pos.X / 60, i.Pos.Y / 60].Value = (int)Player.BlackMen;
-                    //if (i.Player == Player.WhiteMen)
-                    //    board[i.Pos.X / 60, i.Pos.Y / 60].Value = (int)Player.WhiteMen;
-                    board2[i.Pos.Y / 60, i.Pos.X / 60].Value = (int)i.Player;
-                }
-
-                buffer_move_black = MovesJumps.RunBlack(board);
-
-                //for (int i = 0; i < 8; i++)
-                //{
-                //    for (int j = 0; j < 8; j++)
-                //        Console.Write(board2[i, j].Value);
-                //    Console.WriteLine();
-                //}
-
-                foreach (var v in buffer_move_black)
-                {
-                    if (same(v, board2))
-                        goodMove = true;
-                }
-
-                if (goodMove == false)
-                    System.Windows.MessageBox.Show("Niepoprawny ruch.");
-                else
-                {
-
-                    currentMove++;
-                    shownMove++;
-                    TB_MoveNr.Text = "Nr ruchu: " + (shownMove).ToString() + " (bieżący)";
-                    chessBoardState.history.Add(pieces2);
-                }
-            }
-        }
-
-        private void BT_GoBack_Click(object sender, RoutedEventArgs e)
-        {
-            if (shownMove < 1)
-                return;
-
-            shownMove--;
-
-            if (shownMove==0)
-                TB_MoveNr.Text = "Nr ruchu: Ustawianie pionków";
-            else
-                TB_MoveNr.Text = "Nr ruchu: " + shownMove.ToString();
-
-            BT_SaveMove.IsEnabled = false;
-            BT_ImageTest.IsEnabled = false;
-
-            chessBoardState.piecesObservable.Clear();
-            foreach (var p in chessBoardState.history[shownMove])
-                chessBoardState.piecesObservable.Add(p.Copy());
-
-        }
-
-        private void BT_GoForward_Click(object sender, RoutedEventArgs e)
-        {
-            if (shownMove == currentMove)
-                return;
-
-            shownMove++;
-            
-
-            if (shownMove == currentMove)
-            {
-                BT_SaveMove.IsEnabled = true;
-                BT_ImageTest.IsEnabled = true;
-                TB_MoveNr.Text = "Nr ruchu: "+ (shownMove).ToString()+" (bieżący)";
-                return;
-            }
-
-            TB_MoveNr.Text = "Nr ruchu: " + (shownMove).ToString();
-            chessBoardState.piecesObservable.Clear();
-            foreach (var p in chessBoardState.history[shownMove])
-                chessBoardState.piecesObservable.Add(p.Copy());
-        }
-
-        private void Button_Click(object sender, RoutedEventArgs e)
-        {
-            MessageBoxResult result = System.Windows.MessageBox.Show("Czy na pewno chcesz rozpocząć grę od nowa?\nDotychczasowy postęp zostanie utracony", "Checkers Project", MessageBoxButton.YesNo, MessageBoxImage.Question);
-            if (result == MessageBoxResult.Yes)
-            {
-                chessBoardState.history.Clear();
-                currentMove = 0;
-                shownMove = 0;
-                TB_MoveNr.Text = "Nr ruchu: Ustawianie pionków";
-                BT_SaveMove.IsEnabled = true;
-            }
-        }
 
         private bool VerifyMoveJumps(ChessField[,] fields)
         {
@@ -682,18 +372,6 @@ namespace CheckersApplication
             }
 
             return false;
-        }
-
-        private void Button_Click_1(object sender, RoutedEventArgs e)
-        {
-            try
-            {
-                Detect(fileName);
-            }
-            catch (Exception ex)
-            {
-                System.Windows.MessageBox.Show("Nie wczytano obrazka.");
-            }
         }
     }
